Scripts/requirements.txt
requests==2.31.0
pymongo==3.12.0
cassandra-driver==3.28.0
neo4j==5.15.0
pandas==2.0.3